fix(checker): keep context module below arch limit
