+++ crates/tsz-checker/src/context/mod.rs
@@ -59,6 +59,7 @@ mod source_file_symbol_type_cache_scope;
 pub(crate) mod speculation;
 mod strict_mode;
 mod symbol_file_targets;
+mod type_parameter_scope_cache_key;
 pub mod typing_request;
 pub use aliases::*;
 pub(crate) use diagnostic_indices::DiagnosticIndices;
@@ -71,6 +72,7 @@ use std::rc::Rc;
 use std::sync::Arc;
 pub use symbol_file_targets::SymbolFileTargetsOverlay;
 use tsz_common::interner::Atom;
+pub use type_parameter_scope_cache_key::TypeParameterScopeCacheKey;
 pub use typing_request::{ContextualOrigin, FlowIntent, TypingRequest};
 
 use crate::control_flow::FlowGraph;
@@ -96,32 +98,6 @@ use tsz_parser::parser::node::NodeArena;
 /// caller sees the first caller's work.
 pub type CrossFileTypeParamsCache =
     Arc<dashmap::DashMap<(u32, NodeIndex), Vec<tsz_solver::TypeParamInfo>>>;
-
-/// Cache key for type-node results resolved under active generic bindings.
-///
-/// Plain `node_types` entries are keyed only by `NodeIndex`, so they are safe
-/// only when no dynamic type-parameter scope is active. When generic bindings
-/// are present, the same annotation node can resolve to different `TypeId`s.
-/// This key records the active name-to-type binding set in deterministic order
-/// so equivalent scopes share work without letting different generic contexts
-/// alias each other.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct TypeParameterScopeCacheKey(Vec<(String, TypeId)>);
-
-impl TypeParameterScopeCacheKey {
-    #[must_use]
-    pub fn from_scope(scope: &FxHashMap<String, TypeId>) -> Option<Self> {
-        if scope.is_empty() {
-            return None;
-        }
-        let mut entries: Vec<_> = scope
-            .iter()
-            .map(|(name, &type_id)| (name.clone(), type_id))
-            .collect();
-        entries.sort_unstable_by(|left, right| left.0.cmp(&right.0));
-        Some(Self(entries))
-    }
-}
 
 /// Overflow state observed from relation checks that feed assignability
 /// diagnostics.
